refactor lib files
In an effort to simply things, adding new Routes() class to handle uri to controller connection. Adding Routes->add() method to specify the uri to controller connection, rather than using logic to return the correct controller based on the uri.
