+++ web/.htaccess
@@ -4,4 +4,4 @@ RewriteRule ^(.*)$ $1
 
 RewriteCond %{REQUEST_FILENAME} !-d
 RewriteCond %{REQUEST_FILENAME} !-f
-RewriteRule ^(.*)$ /index.php?q=$1 [L,QSA]
+RewriteRule ^(.*)$ /index.php?uri=$1 [L,QSA]
