+++ web/index.php
@@ -1,23 +1,16 @@
 <?php
 require "../autoload.php";
 
-use Lib\Bootstrap\Bootstrap;
-use Lib\Db\Db;
+use Lib\Route;
 
-// grabs URI
-if (!isset($_GET["q"])) {
-    $uri = 'index';
-} else {
-    $uri = $_GET["q"];
-}
-
-// initialize database
-$db = Db::getInstance();
-
-// routing
-$bootstrap = new Bootstrap($uri);
-$action = $bootstrap->request->action."Action";
-$data = $bootstrap->controller->$action();
+$route = new Route();
+/*
+* add any new routes below
+*/
+//////////////////////////
+$route->add("/", "index", "index");
 
 
-require_once(VIEWS."/layout.html");
+//////////////////////////
+$uri = $_GET["uri"] ?? '';
+$route->get($uri);
